Fix: Stage detection and initial Stage 2 data for smooth UI transitions
Fixed two issues causing UI glitches when transitioning to Stage 2:

Issue 1: Type Mismatch in Stage Detection
- JSON stores stage as string "1" or "2"
- Comparison was $currentStage === 1 (strict int comparison)
- String "2" !== int 1 was always false
- Fix: Cast to int: (int) ($data['stage'] ?? 1)

Issue 2: Missing Data in Initial Stage 2 Update
- When Stage 2 started, updateProgress() only set stage: 2
- Missing: total_processed, total_expected, created, updated, failed
- UI would briefly show 0/0 or undefined values
- Fix: Include all expected fields in initial Stage 2 update

Changes:
- ImportProgress.php: Cast stage to int for proper comparison
- SyncHistoricalOrdersJob.php: Include complete progress_data when Stage 2 starts
  - Shows "Starting import of X orders..."
  - Initializes all counters to 0
  - Includes total_expected from Stage 1

Result:
- Clean transition: Stage 1 complete → Stage 2 starts at 10% with 0/182,931
- No more grey badge or empty progress bar
- Stage 2 badge immediately shows blue with pulse animation
- Smooth, professional UX

+++ app/Livewire/Settings/ImportProgress.php
@@ -101,8 +101,8 @@ class ImportProgress extends Component
         if ($activeSync->progress_data) {
             $data = $activeSync->progress_data;
 
-            // Determine current stage
-            $this->currentStage = $data['stage'] ?? 1;
+            // Determine current stage (cast to int for strict comparison)
+            $this->currentStage = (int) ($data['stage'] ?? 1);
 
             if ($this->currentStage === 1) {
                 // Stage 1: Streaming order IDs
